monolith/requirements.txt
scikit-learn==0.23.2
cloudmesh-common==4.3.29
pillow==8.0.1